imginimg/requirements.txt
boto3
botocore==1.4.72
chalice==0.4.0
click==6.2
cv2wrap==1.0
dlib==19.1.0
docutils==0.12
jmespath==0.9.0
numpy==1.11.2
opencv-python==3.1.0.3
Pillow==3.4.2
pkg-resources==0.0.0
py-bing-search==0.2.6
python-dateutil==2.6.0
requests==2.12.0
six==1.10.0
typing==3.5.1.0
virtualenv==15.0.3
